webtrinity/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Login>=0.6
Werkzeug>=2.2
psycopg2-binary>=2.9
gunicorn>=20.1
python-dotenv>=1.0

